Move device code back to backends
