+++ taichi/backends/device.cpp
@@ -1,7 +1,7 @@
 #include <taichi/backends/device.h>
 
 #if TI_WITH_VULKAN
-#include <taichi/runtime/vulkan/vulkan_device.h>
+#include <taichi/backends/vulkan/vulkan_device.h>
 #include <taichi/backends/interop/vulkan_cpu_interop.h>
 #if TI_WITH_LLVM
 #include <taichi/backends/cpu/cpu_device.h>
