+++ taichi/backends/vulkan/vulkan_device_creator.h
@@ -4,7 +4,7 @@
 #define VK_USE_PLATFORM_WIN32_KHR 1
 #endif
 
-#include "taichi/runtime/vulkan/vulkan_common.h"
+#include "taichi/backends/vulkan/vulkan_common.h"
 
 #include <taichi/backends/device.h>
