+++ taichi/runtime/program_impls/CMakeLists.txt
@@ -1,0 +1,3 @@
+# ./taichi/runtime/program_impls/CMakeLists.txt
+
+add_subdirectory(vulkan)
