+++ taichi/runtime/program_impls/vulkan/CMakeLists.txt
@@ -1,0 +1,18 @@
+# ./taichi/runtime/program_impls/vulkan/CMakeLists.txt
+
+add_library(vulkan_program_impl)
+target_sources(vulkan_program_impl
+  PRIVATE
+    vulkan_program.cpp
+  )
+
+target_include_directories(vulkan_program_impl
+  PRIVATE
+    ${PROJECT_SOURCE_DIR}
+    ${PROJECT_SOURCE_DIR}/external/eigen
+    ${PROJECT_SOURCE_DIR}/external/spdlog/include
+    ${PROJECT_SOURCE_DIR}/external/SPIRV-Tools/include
+  )
+
+target_link_libraries(vulkan_program_impl PRIVATE vulkan_rhi)
+target_link_libraries(vulkan_program_impl PRIVATE gfx_runtime)
