+++ taichi/ui/backends/vulkan/renderables/mesh.cpp
@@ -1,7 +1,7 @@
 #include "mesh.h"
 
 #include "taichi/ui/utils/utils.h"
-#include "taichi/runtime/vulkan/vulkan_device.h"
+#include "taichi/backends/vulkan/vulkan_device.h"
 
 TI_UI_NAMESPACE_BEGIN
