+++ tests/cpp/aot/aot_save_load_test.cpp
@@ -10,10 +10,10 @@
 #include "taichi/runtime/gfx/aot_module_loader_impl.h"
 #ifdef TI_WITH_VULKAN
 #include "taichi/backends/device.h"
-#include "taichi/runtime/vulkan/vulkan_device.h"
-#include "taichi/runtime/vulkan/vulkan_device_creator.h"
-#include "taichi/runtime/vulkan/vulkan_loader.h"
-#include "taichi/runtime/vulkan/vulkan_utils.h"
+#include "taichi/backends/vulkan/vulkan_device.h"
+#include "taichi/backends/vulkan/vulkan_device_creator.h"
+#include "taichi/backends/vulkan/vulkan_loader.h"
+#include "taichi/backends/vulkan/vulkan_utils.h"
 #endif
 
 using namespace taichi;
