+++ tests/cpp/aot/runtime_test.cpp
@@ -7,10 +7,10 @@
 #include "taichi/runtime/gfx/runtime.h"
 #ifdef TI_WITH_VULKAN
 #include "taichi/backends/device.h"
-#include "taichi/runtime/vulkan/vulkan_device.h"
-#include "taichi/runtime/vulkan/vulkan_device_creator.h"
-#include "taichi/runtime/vulkan/vulkan_loader.h"
-#include "taichi/runtime/vulkan/vulkan_utils.h"
+#include "taichi/backends/vulkan/vulkan_device.h"
+#include "taichi/backends/vulkan/vulkan_device_creator.h"
+#include "taichi/backends/vulkan/vulkan_loader.h"
+#include "taichi/backends/vulkan/vulkan_utils.h"
 #endif
 
 using namespace taichi;
